# backend/openapischema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Document Analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/podcasts/generate/from-document/{document_id}": {
      "post": {
        "tags": ["Podcasts"],
        "summary": "Generate a podcast from a document",
        "description": "Generates an audio podcast from the content of a specified document. The generation process is synchronous. Returns the details of the newly created podcast, including its audio URL and transcript.",
        "operationId": "generate_podcast_from_document_api_v1_podcasts_generate_from_document__document_id__post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the document to generate the podcast from.",
              "title": "Document Id"
            },
            "description": "The ID of the document to generate the podcast from."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PodcastGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/podcasts/generate/from-recommendation/{recommendation_id}": {
      "post": {
        "tags": ["Podcasts"],
        "summary": "Generate a podcast from a recommendation",
        "description": "Generates an audio podcast from the content of a specified recommendation. The generation process is synchronous. Returns the details of the newly created podcast, including its audio URL and transcript.",
        "operationId": "generate_podcast_from_recommendation_api_v1_podcasts_generate_from_recommendation__recommendation_id__post",
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the recommendation to generate the podcast from.",
              "title": "Recommendation Id"
            },
            "description": "The ID of the recommendation to generate the podcast from."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PodcastGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/podcasts/{podcast_id}": {
      "get": {
        "tags": ["Podcasts"],
        "summary": "Retrieve a single podcast by ID",
        "description": "Fetches a single podcast's details using its unique identifier. Returns a 404 error if the podcast is not found.",
        "operationId": "get_podcast_api_v1_podcasts__podcast_id__get",
        "parameters": [
          {
            "name": "podcast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the podcast to retrieve.",
              "title": "Podcast Id"
            },
            "description": "The unique identifier of the podcast to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/podcasts/generate/from-collection/{collection_id}": {
      "post": {
        "tags": ["Podcasts"],
        "summary": "Generate a podcast from a collection",
        "description": "Generates an audio podcast from the aggregated content of documents within a specified collection. The generation process is synchronous. Returns the details of the newly created podcast, including its audio URL and transcript.",
        "operationId": "generate_podcast_from_collection_api_v1_podcasts_generate_from_collection__collection_id__post",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the collection to generate the podcast from.",
              "title": "Collection Id"
            },
            "description": "The ID of the collection to generate the podcast from."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PodcastGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PodcastGenerateRequest": {
        "properties": {
          "include_insights": {
            "type": "boolean",
            "title": "Include Insights",
            "description": "Flag to include insights in the podcast summary.",
            "default": false
          },
          "min_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Duration Seconds",
            "description": "Minimum desired podcast duration in seconds (default: 120s = 2min).",
            "default": 120
          },
          "max_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Duration Seconds",
            "description": "Maximum desired podcast duration in seconds (default: 240s = 4min).",
            "default": 240
          }
        },
        "type": "object",
        "title": "PodcastGenerateRequest",
        "description": "Pydantic model for the request body to generate a podcast."
      },
      "PodcastInDB": {
        "properties": {
          "sourceType": {
            "type": "string",
            "title": "Sourcetype",
            "description": "The type of the source ('document' or 'recommendation')."
          },
          "sourceId": {
            "type": "string",
            "title": "Sourceid",
            "description": "The ID of the source entity."
          },
          "transcript": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PodcastSegment"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript",
            "description": "A list of podcast segments, representing the full transcript."
          },
          "podcastId": {
            "type": "string",
            "title": "Podcastid",
            "description": "The unique identifier for the podcast."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "The current status of the generation job."
          },
          "audioUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audiourl",
            "description": "The URL where the generated audio file is stored."
          },
          "durationSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationseconds",
            "description": "The length of the audio in seconds."
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat",
            "description": "The timestamp when the podcast was generated."
          },
          "shortDescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortdescription",
            "description": "A two-sentence description of what the podcast is all about."
          }
        },
        "type": "object",
        "required": [
          "sourceType",
          "sourceId",
          "podcastId",
          "status",
          "generatedAt"
        ],
        "title": "PodcastInDB",
        "description": "Pydantic model representing a podcast as stored in the database.\nThis model is used for API responses."
      },
      "PodcastSegment": {
        "properties": {
          "speaker": {
            "type": "string",
            "title": "Speaker",
            "description": "The speaker for this segment (e.g., 'HOST', 'GUEST')."
          },
          "dialogue": {
            "type": "string",
            "title": "Dialogue",
            "description": "The dialogue for this segment."
          },
          "words": {
            "type": "integer",
            "title": "Words",
            "description": "The estimated word count for this segment."
          },
          "order": {
            "type": "integer",
            "title": "Order",
            "description": "The order of this segment in the overall script."
          }
        },
        "type": "object",
        "required": ["speaker", "dialogue", "words", "order"],
        "title": "PodcastSegment",
        "description": "Represents a single segment of the podcast script."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}